+++ crates/beskid_analysis/src/format/types_emit.rs
@@ -44,11 +44,13 @@ impl Emit for PrimitiveType {
             PrimitiveType::I32 => "i32",
             PrimitiveType::I64 => "i64",
             PrimitiveType::U8 => "u8",
+            PrimitiveType::Pointer => "pointer",
             PrimitiveType::Word => "word",
             PrimitiveType::F64 => "f64",
             PrimitiveType::Char => "char",
             PrimitiveType::String => "string",
             PrimitiveType::Unit => "unit",
+            PrimitiveType::Never => "never",
         })?;
         Ok(())
     }