+++ crates/beskid_analysis/src/hir/lowering/types.rs
@@ -4,8 +4,8 @@ use crate::hir::{
 };
 use crate::syntax::{self, Spanned};
 
-use super::items::lower_attributes;
 use super::Lowerable;
+use super::items::lower_attributes;
 
 impl Lowerable for Spanned<syntax::Type> {
     type Output = Spanned<HirType>;
@@ -157,11 +157,13 @@ impl Lowerable for Spanned<syntax::PrimitiveType> {
                 syntax::PrimitiveType::I32 => HirPrimitiveType::I32,
                 syntax::PrimitiveType::I64 => HirPrimitiveType::I64,
                 syntax::PrimitiveType::U8 => HirPrimitiveType::U8,
+                syntax::PrimitiveType::Pointer => HirPrimitiveType::Word,
                 syntax::PrimitiveType::Word => HirPrimitiveType::Word,
                 syntax::PrimitiveType::F64 => HirPrimitiveType::F64,
                 syntax::PrimitiveType::Char => HirPrimitiveType::Char,
                 syntax::PrimitiveType::String => HirPrimitiveType::String,
                 syntax::PrimitiveType::Unit => HirPrimitiveType::Unit,
+                syntax::PrimitiveType::Never => HirPrimitiveType::Never,
             },
             self.span,
         )